chore: improve docs, move functions around #136
